auth/auth/tests/requirements.txt
pydantic
aiohttp
elasticsearch[async]==7.9.1
httptools==0.5.0
redis==4.4.0
Flask==2.2.2
SQLAlchemy==1.4.45
psycopg2==2.9.5
flask_sqlalchemy==3.0.2
pyjwt==2.6.0
flask_jwt_extended==4.4.4
wtforms==3.0.1
pytest-asyncio==0.12.0
faker==0.7.4
requests==2.28.1
flasgger==0.9.5
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
CacheControl==0.12.11
